requirements.txt
chromadb==0.3.25
langchain==0.0.183
python-dotenv==1.0.0
streamlit==1.22.0
